Move the tests into MoonshotChatModelFunctionCallingIT
